AI/requirements.txt
# Core RAG Pipeline Dependencies
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-groq>=0.1.0
langchain-huggingface>=0.3.0
langchainhub>=0.1.0

# Vector Database
chromadb>=1.0.0

# Embeddings
sentence-transformers>=2.0.0
huggingface-hub>=0.20.0

# Text Processing
tiktoken>=0.5.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# HTTP Client
httpx>=0.24.0
requests>=2.30.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0

# Utilities
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
